Demo1 Figuring out
Figured out how to instantiate new cars that point and go to a waypoint.  Starting working on a new scene that I hope will be a demo.  I need to figure out how to make a waypoint when a blue cube is made then have that waypoint be accessible to the other waypoints